Change focal length to float,

+++ database/migrations/2019_07_30_170211_create_eyepieces_table.php
@@ -17,9 +17,9 @@ class CreateEyepiecesTable extends Migration
             function (Blueprint $table) {
                 $table->bigIncrements('id');
                 $table->string('name', 255);
-                $table->unsignedInteger('focalLength');
+                $table->float('focalLength');
                 $table->unsignedInteger('apparentFOV');
-                $table->unsignedInteger('maxFocalLength')->nullable();
+                $table->float('maxFocalLength')->nullable();
                 $table->unsignedInteger('observer_id');
                 $table->boolean('active')->default(true);
                 $table->unsignedInteger('observations')->default(0);
